Rounds detection of the location of center of pierce-only holes
Fusion 360 is known to create multiple, partial circular operations (G3) for
a single hole. Sometimes, the center of the sequence of operations can differ
slightly and the pierce-only detection would create several pierce-only
operations for the same hole. By rounding down the location of the center,
we can now detect the same hole and avoid multiple spurious operations.
The detection limit is less than one mm or less than one tenth of an inch.